First implementation of new map
There are still a lot of issues, such as the lack of all the old layers,
left over code, and a few missing features (move to for example) But
this is a start.

Note that caching methods should not be bothered with for the moment, as
a cache directory and expiration date have not been coded.