fixed a NullPointerException on startup
When Amidst is started without settings file, it creates a settings file with default values. One of the default values was null, which is not allowed as a value in a preference store. I changed it to use the empty String instead. I also added some checks to ensure this does not happen again.