moved devtools to the test source folder
The purpose of the devtools source folder was to have a place to put code that should not be included in the released jar file. However, this requirement is also met by the test source folder. The devtools source folder was a relict from before we used maven.

I also externalized the devtool settings to the properties file `src/test/resources/amidst/devtools/devtools.properties`. This file is ignored by git, however there is a commited template file. To run the devtools, copy the template and populate it with the correct values. The template also contains documentation about how the values are used.

Added tests to execute each devtool. Removed all main methods from the devtools. The tests are ignored by default since they serve as a main methods.