changed worker threads to be no daemon threads
If a thread is a daemon thread, it does not keep the JVM alive. This means, that if all non-daemon threads are terminated, the JVM will just terminate all daemon threads and exit. This is a bad thing, if the daemon thread is currently writing to a file, because this will corrupt the file. Since the worker threads are used for IO, they should not be daemon threads.