converted the CrashWindow into a singleton
Before, when there was a bug in the drawing loop this would have caused Amidst to open up to 50 crash windows per second (Amidst is capped at 50 FPS), which is really bad. Also, it is very likely that such bugs would not be reported, because the bug report cannot be copied.

Now, only one CrashWindow is opened and the displayed error message is updated when more crashes occur.