updated the README file
Placed the current version of the preliminary Forum Entry to the README file. If more changes should be made to the Forum Entry, please do it in the wiki. We will update the README again as soon as the Forum Entry was created.

This was done to fix the broken links. Also, it was commited directly to master, because there is no way this will break the build.