switched windows to use the default Java look and feel
This is because the Windows look and feel does not display checkboxes for menu items that have an icon. Also, the majority of the GUI is drawn by an own painting method, which does not rely on the look and feel. Thus, the decision to switch it should not have a great impact on the usability. Another advantage of this change is, that now all systems use the same look and feel, which should lower the chance of bugs happening on only one type of system.