After loading registry, stop datafixer threads
The threads that register all of the datafixers take up a TON of CPU usage for a few minutes after loading the Registry class, and it doesn't help us in any way. We can just stop them a bit after they are created by getting the ExecutorService running them. I couldn't find a way to load the class without the threads being created, so I just did this instead.