Fix Fragment loading ceasing after switching Seeds
This commit fixes an issue where Fragments would stop being loaded if
the current world was switched while the fragment workers were busy.

When a new world is viewed, ViewerFacade#dispose() is called. First, the
fragmentManager is cleared(), removing any pending fragments. Then, the
fragment Thread-Pool is shut down without waiting for its jobs to
finish. The FragmentQueueProcessor is responsible for dispatching work
to the fragment workers. It parks if the queue contains items while all
workers are busy, waiting for a worker to unpark it after finishing its
work. If no worker finishes its work between the call to
FragmentManager#clear() and the fragment pool being shutdownNow()'d,
then the FragmentQueueProcessor would never be unparked, resulting in no
fragments being loaded. Well, there is a timeout set for parking, but
it's at 11.5 days...

This issue was not observed in practice until I added the *glacially*
slow infdev support, which hit that issue quite consistently.

This commit fixes the issue by unparking the fragment loader Thread when
a runnable is assigned to it and the thread has started. Setting the
timeout lower also worked, but didn't really adress the bug.