Support early betas
This commit implements the functionality of the support_early_betas
branch from 2016. The code is mostly a translation into the current set
of classes / builders, though I needed to add a bit of infrastructure.
I didn't add oceans yet, as they are not a proper biome in these
versions.

I expanded the clazz package to allow matching int constants, because
that functionality was used by support_early_betas. I had to increase
the MAXIMUM_CLASS_BYTES to match the World class in a few beta versions.