Implement oceans in beta terrain
This commit enables displaying oceans in beta versions. I was able to
largely solve the performance issues by implementing a custom noise
generator that only generates at ocean level and by only generating the
most significant octaves.

We are now running part of Minecraft's chunk generation, the
`shapeChunk` method. This uses several PerlinOctaveNoises to determine
for each block in a chunk if it should be air, stone, water, or ice.
As was discovered on the support_early_betas branch, just doing this
would be *very* slow.

This commit employs two major optimizations to get good performance,
both targeting Minecraft's 3D interpolation noise. This noise is
responsible for scrambling the terrain generated by the depth and biome
noises and is vital for accurately showing oceans. The vanilla
implementation is wasteful in two ways: It generates noise for Y values
which are not relevant for determining the blocks at Y=63 and it uses 16
octaves of nosie. My optimized implementation skips irrelevant Y values
and allows us to limit the number of octaves to only the n most
significant ones. This setting is controlled via the OCEAN_PRECISION
constant, I set it to 4 as that provides good accuracy while still being
fast.

This involved implementing my own PerlinNoise and PerlinOctaveNoise
classes. The code is mostly just a straightforward implementation of
Perlin's algorithm while conforming to the interface Minecraft expects.
I actually had to implement a bug, as Minecraft's PerlinNoise
implementation is slightly broken.

This commit adds a new dependency, Byte Buddy, for generating two
classes at runtime:

In order to access an OverworldLevelSource, we need to isntantiate an
OverworldDimension. Until b1.6, OverworldDimension was simply the
non-abstract base of all other dimensions. However, it became abstract
in b1.6 with DimensionOverworld now being an empty class inheriting from
it. I did not know of a way to match this DimensionOverworld, so I am
using Byte Buddy to create a class that inherits from Dimension at
runtime.

Byte Buddy is used to generate a subclass of Minecraft's
PerlinOctaveNoise that delegates to our optimized implementation. These
optimizations are vital to supporting accurate oceans at a reasonable
performance and this was the best way I knew to add them.