Match overworld dimension instead of generating it
This commit changes the BetaMinecraftInterface to instantiate the real
DimensionOverworld class, rather than generating an inheriting class
with ByteBuddy. To accomplish this, I expanded the ClassTranslator to
allow access to previously matched classes. This now lets us specify
"empty class that inherits from DimensionBase", thus allowing us to
match DimensionOverworld.

I had to change ClassTranslator.translations to a List, so that
iteration order would be guaranteed. I don't see why this field was a
map before, other than the need to store tuples.